backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-openai==0.3.0
openai==1.58.1
supabase==2.11.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.4
httpx==0.28.1
PyPDF2==3.0.1
python-docx==1.1.2
slowapi==0.1.9
